feat():validations to category router only admin access
and take in count field isVerified to allow user to create offers or modify it, and other validations

+++ routers/category.js
@@ -2,6 +2,7 @@ const express = require('express');
 const errorHandler = require('../helpers/errorHandler');
 const {Category} = require('../models/category');
 const router = express.Router();
+const jwt = require('jsonwebtoken');
 
 
 /**
@@ -29,19 +30,25 @@ router.get('/:id',async(req,res)=>{
 
 
 router.post('/', async(req,res)=>{
+
+    const isAdmin = jwt.decode(req.headers.authorization.split(' ')[1]).isAdmin
+    if(!isAdmin) return res.status(400).send('Autorización requerida.');
+    
     let category = new Category({
         name: req.body.name,
         description:req.body.description,
         icon:null
     });
-
+    
     category = await category.save();
-
     if(!category) return res.status(404).send('The category cannot be created!')
     res.send(category);
 });
 
 router.put('/:id', async(req,res)=>{
+    const isAdmin = jwt.decode(req.headers.authorization.split(' ')[1]).isAdmin
+    if(!isAdmin) return res.status(400).send('Autorización requerida.');
+    
     const category  = await Category.findByIdAndUpdate(req.params.id,{
         name: req.body.name,
         description:req.body.description,
@@ -56,6 +63,9 @@ router.put('/:id', async(req,res)=>{
 
 router.delete('/:id',async(req,res)=>{
 
+    const isAdmin = jwt.decode(req.headers.authorization.split(' ')[1]).isAdmin
+    if(!isAdmin) return res.status(400).send('Autorización requerida.');
+    
     Category.findByIdAndDelete(req.params.id).then(category =>{
         if(category){
             return res.status(200).json({success:true,message:'The category was deleted.'});
